Move score store
